chore: consolidates project detail content into one file per project and update site content
