+++ src/components/Actors/Actors.jsx
@@ -1,24 +1,23 @@
 import React from 'react';
 import { Divider } from '../Divider';
 
-const Actors = () => {
+const Actors = ({ content }) => {
   return (
     <div>
       <h1 className="h1 text-primary-yellow mb-0 desktop:mb-[0.625rem] ml-0">
-        Actors
+        {content.title}
       </h1>
       <div className="flex gap-[2.25rem] tablet:gap-[3rem] desktop:gap-[4rem] flex-col desktop:flex-row">
         <div className="mt-[2.125rem] tablet:mt-[3.4375rem]">
           <div className="mb-[2.25rem] desktop:mb-0">
             <div className="flex gap-6 mb-1">
               <img src="/img/actors-index-1.svg" alt="1" />
-              <h2 className="h2 text-primary-yellow flex items-center">
-                Wallets
+              <h2 className="h2-caps text-primary-yellow flex items-center">
+                {content.actors[0].title}
               </h2>
             </div>
             <p className="copy text-primary-yellow flex ml-[4.375rem] desktop:ml-0">
-              Wallets act as agents for individuals or institutions by
-              facilitating exchanges with PFIs
+              {content.actors[0].description}
             </p>
           </div>
           <div className="mt-[3rem] mb-[2.25rem] desktop:block hidden">
@@ -27,11 +26,12 @@ const Actors = () => {
           <div className="mb-[2.25rem] desktop:mb-0">
             <div className="flex gap-6 mb-1">
               <img src="/img/actors-index-2.svg" alt="2" />
-              <h2 className="h2 text-primary-yellow flex items-center">PFIS</h2>
+              <h2 className="h2-caps text-primary-yellow flex items-center">
+                {content.actors[1].title}
+              </h2>
             </div>
             <p className="copy text-primary-yellow flex ml-[4.375rem] desktop:ml-0">
-              Participating financial institutions that offer liquidity services
-              on the tbDEX network
+              {content.actors[1].description}
             </p>
           </div>
           <div className="mt-[3rem] mb-[2.25rem] desktop:block hidden">
@@ -40,13 +40,12 @@ const Actors = () => {
           <div>
             <div className="flex gap-6 mb-1">
               <img src="/img/actors-index-3.svg" alt="3" />
-              <h2 className="h2 text-primary-yellow flex items-center">
-                CREDENTIAL ISSUER
+              <h2 className="h2-caps text-primary-yellow flex items-center">
+                {content.actors[2].title}
               </h2>
             </div>
             <p className="copy text-primary-yellow flex ml-[4.375rem] desktop:ml-0">
-              Organizations or individuals (by means of their wallet) who serve
-              as a source of verifiable credentials
+              {content.actors[2].description}
             </p>
           </div>
         </div>
