+++ src/components/Actors/Actors.stories.jsx
@@ -1,5 +1,6 @@
 import { Actors } from '.';
 import React from 'react';
+import { content } from '@site/src/content/project-content-tbdex.js';
 export default {
   title: 'Actors',
   component: Actors,
@@ -7,7 +8,7 @@ export default {
 
 const TemplateDark = () => (
   <div className="dark bg-primary-black">
-    <Actors />
+    <Actors content={content.actors} />
   </div>
 );
