--- src/content/components-web5.js
@@ -1,41 +1,0 @@
-/* eslint-disable quotes */
-export const content = {
-  growToFit: true,
-  title: 'Components',
-  projects: [
-    {
-      icon: '/img/component-icon.svg',
-      title: 'Decentralized Identifiers',
-      description:
-        'An implementation of DIF’s emerging decentralized identifiers standard.',
-      textButton: 'View Component',
-      url: 'https://identity.foundation/ion/',
-      isExternalLink: true,
-    },
-    {
-      icon: '/img/component-icon.svg',
-      title: 'Decentralized Web Node',
-      description:
-        "An implementation of DIF's emerging decentralized personal datastore standard.",
-      textButton: 'View Component',
-      url: '/projects/dwn-sdk-js/README',
-    },
-
-    {
-      icon: '/img/component-icon.svg',
-      title: 'Self-Sovereign Identity Service',
-      description:
-        'An in-a-box service that handles the full Verifiable Credentials lifecycle, including issuance, verification, revocation, and more.',
-      textButton: 'View Component',
-      url: '/projects/ssi-service/README',
-    },
-    {
-      icon: '/img/component-icon.svg',
-      title: 'Self-Sovereign Identity SDK',
-      description:
-        'Standards-based primitives for using Decentralized Identifiers and Verifiable Credentials.',
-      textButton: 'View Component',
-      url: '/projects/ssi-sdk/README',
-    },
-  ],
-};
