+++ src/pages/projects-index.mdx
@@ -38,9 +38,5 @@ In particular, we want to ensure the New Contributor Experience is as smooth as 
 - Clone and build the project
 - Run the tests
 - Join the conversation in Discussions and Issues
-<!--stackedit_data:
-eyJoaXN0b3J5IjpbLTE2NjA3ODA4ODksLTE3NDQ2MTU1MjBdfQ
-==
--->
 
 </div>
